se resolvio un conflicto que se probocaba al traer varios creditos en el componente cardstop
