+++ front/src/components/Context/SidebarAdmin.jsx
@@ -21,7 +21,7 @@ const SidebarAdmin = () => {
               </Button>
             </li>
             <li>
-              <Button as={Link} to="cargarCreditos" className="block p-2 transition rounded w-100 hover:bg-blue-700">
+              <Button as={Link} to="asingCredit" className="block p-2 transition rounded w-100 hover:bg-blue-700">
                 Cargar Creditos
               </Button>
             </li>
